+++ LPC/LPC_and_LFCC.h
@@ -2,7 +2,7 @@
 #define _LPC_and_LFCC_h_
 /* LPC_and_LFCC.h
  *
- * Copyright (C) 1994-2011, 2015 David Weenink
+ * Copyright (C) 1994-2017 David Weenink
  *
  * This code is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -27,9 +27,9 @@
 #include "LPC.h"
 #include "LFCC.h"
 
-autoLFCC LPC_to_LFCC (LPC me, long numberOfCoefficients);
+autoLFCC LPC_to_LFCC (LPC me, integer numberOfCoefficients);
 
-autoLPC LFCC_to_LPC (LFCC me, long numberOfCoefficients);
+autoLPC LFCC_to_LPC (LFCC me, integer numberOfCoefficients);
 
 void LPC_Frame_into_CC_Frame (LPC_Frame me, CC_Frame thee);
 /*
