+++ LPC/LPC_to_Spectrogram.cpp
@@ -26,7 +26,7 @@
 autoSpectrogram LPC_to_Spectrogram (LPC me, double dfMin, double bandwidthReduction, double deEmphasisFrequency) {
 	try {
 		double samplingFrequency = 1.0 / my samplingPeriod;
-		long nfft = 2;
+		integer nfft = 2;
 		if (dfMin <= 0) {
 			nfft = 512; dfMin = samplingFrequency / nfft;
 		}
@@ -37,10 +37,10 @@ autoSpectrogram LPC_to_Spectrogram (LPC me, double dfMin, double bandwidthReduct
 
 		autoSpectrogram thee = Spectrogram_create (my xmin, my xmax, my nx, my dx, my x1, 0.0, samplingFrequency / 2.0, nfft / 2 + 1, freqStep, 0.0);
 
-		for (long i = 1; i <= my nx; i ++) {
+		for (integer i = 1; i <= my nx; i ++) {
 			double t = Sampled_indexToX (me, i);
 			autoSpectrum spec = LPC_to_Spectrum (me, t, dfMin, bandwidthReduction, deEmphasisFrequency);
-			for (long j = 1; j <= spec -> nx; j ++) {
+			for (integer j = 1; j <= spec -> nx; j ++) {
 				double re = spec -> z [1] [j], im = spec -> z [2] [j];
 				thy z [j] [i] =  re * re + im * im;
 			}
